More accurate speed calculation
Rather than doing some stateful crazy dance, we keep track of the times it took to switch phase for the last 1000us.  We then average those.  If our current phase change has gone over that average then we include it in the calculation.  This gets us a much more stable speed estimate, and a much faster drop to 0 when the motor stops.  That in turn makes the pid much more stable.  Unfortunately, we're still stuck not detecing zero crossings just now.